feat: Move `getEmojiByCountry` to country model
BREAKING CHANGE: Please replace your import paths
Before:
```
import { getEmojiByCountry } from "cozy-client/dist/models/document/emojiCountry"
```
After:
```
import { getEmojiByCountry } from "cozy-client/dist/models/country/countries"
```
