+++ packages/cozy-client/src/models/country/countries.spec.js
@@ -6,6 +6,7 @@ import {
   getCountryNameByCodeISO,
   getEmojiByCountry
 } from './countries'
+import logger from '../../logger'
 
 describe('checkCountryCode', () => {
   it('should return true for valid country codes', () => {
@@ -120,5 +121,37 @@ describe('getAllNationalities', () => {
     expect(nationalities).toEqual(
       expect.arrayContaining(['Français', 'Américain'])
     )
+  })
+})
+
+describe('getEmojiByCountry', () => {
+  beforeEach(() => {
+    jest.spyOn(logger, 'error').mockImplementation(() => jest.fn())
+  })
+  afterEach(() => {
+    logger.error.mockRestore()
+  })
+
+  it('should return French flag', () => {
+    const res = getEmojiByCountry('fr')
+
+    expect(res).toBe('🇫🇷')
+  })
+  it('should return Belgian flag', () => {
+    const res = getEmojiByCountry('be')
+
+    expect(res).toBe('🇧🇪')
+  })
+  it('should return "null" if the country is not ISO 3166-1 alpha-2 string', () => {
+    const res = getEmojiByCountry('fra')
+
+    expect(logger.error).toHaveBeenCalledTimes(1)
+    expect(res).toBe(null)
+  })
+  it('should return "null" if the country is undefined', () => {
+    const res = getEmojiByCountry(undefined)
+
+    expect(logger.error).not.toHaveBeenCalled()
+    expect(res).toBe(null)
   })
 })
